Bare code move from app-vanitygen
