+++ packages/app-accounts/package.json
@@ -12,7 +12,14 @@
   "dependencies": {
     "@babel/runtime": "^7.2.0",
     "@polkadot/ui-app": "^0.22.18",
+    "@types/yargs": "^12.0.5",
     "@types/file-saver": "^2.0.0",
-    "file-saver": "^2.0.0"
+    "babel-plugin-module-resolver": "^3.1.1",
+    "file-saver": "^2.0.0",
+    "libsodium-wrappers": "^0.7.3",
+    "yargs": "^12.0.5"
+  },
+  "optionalDependencies": {
+    "sodium": "^3.0.2"
   }
 }
