+++ packages/app-accounts/src/Vanity/Match.tsx
@@ -1,11 +1,10 @@
-// Copyright 2017-2019 @polkadot/app-vanitygen authors & contributors
+// Copyright 2017-2019 @polkadot/app-accounts authors & contributors
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
 import { BareProps } from '@polkadot/ui-app/types';
 
 import React from 'react';
-
 import { Button, CopyButton, IdentityIcon } from '@polkadot/ui-app/index';
 import { u8aToHex } from '@polkadot/util';
