+++ packages/app-accounts/src/Vanity/index.tsx
@@ -1,20 +1,20 @@
-// Copyright 2017-2019 @polkadot/app-vanitygen authors & contributors
+// Copyright 2017-2019 @polkadot/app-accounts authors & contributors
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
 import { I18nProps } from '@polkadot/ui-app/types';
 import { ActionStatus } from '@polkadot/ui-app/Status/types';
-import { Generator$Matches, Generator$Result } from './generator/types';
+import { Generator$Matches, Generator$Result } from '../vanitygen/types';
 
 import './index.css';
 
 import React from 'react';
 import { Button, Dropdown, Input, Static } from '@polkadot/ui-app/index';
 
+import generator from '../vanitygen';
+import matchRegex from '../vanitygen/regex';
+import generatorSort from '../vanitygen/sort';
 import Match from './Match';
-import generator from './generator';
-import matchRegex from './generator/regex';
-import generatorSort from './generator/sort';
 import translate from './translate';
 
 type Props = I18nProps & {
