+++ packages/app-accounts/src/vanitygen/cli.ts
@@ -1,5 +1,5 @@
 #!/usr/bin/env node
-// Copyright 2017-2019 @polkadot/app-vanitygen authors & contributors
+// Copyright 2017-2019 @polkadot/app-accounts authors & contributors
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
