+++ packages/app-accounts/src/vanitygen/index.ts
@@ -1,4 +1,4 @@
-// Copyright 2017-2019 @polkadot/app-vanitygen authors & contributors
+// Copyright 2017-2019 @polkadot/app-accounts authors & contributors
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
