--- packages/app-vanitygen/README.md
@@ -1,1 +1,0 @@
-# @polkadot/app-vanitygen
