--- packages/app-vanitygen/package.json
@@ -1,23 +1,0 @@
-{
-  "name": "@polkadot/app-vanitygen",
-  "version": "0.22.18",
-  "main": "index.js",
-  "repository": "https://github.com/polkadot-js/apps.git",
-  "author": "Jaco Greeff <jacogr@gmail.com>",
-  "maintainers": [
-    "Jaco Greeff <jacogr@gmail.com>"
-  ],
-  "contributors": [],
-  "license": "Apache-2.0",
-  "dependencies": {
-    "@babel/runtime": "^7.2.0",
-    "@polkadot/ui-app": "^0.22.18",
-    "@types/yargs": "^12.0.5",
-    "babel-plugin-module-resolver": "^3.1.1",
-    "libsodium-wrappers": "^0.7.3",
-    "yargs": "^12.0.5"
-  },
-  "optionalDependencies": {
-    "sodium": "^3.0.2"
-  }
-}
