+++ tsconfig.json
@@ -14,7 +14,6 @@
       "@polkadot/app-storage/*": [ "packages/app-storage/src/*" ],
       "@polkadot/app-toolbox/*": [ "packages/app-toolbox/src/*" ],
       "@polkadot/app-transfer/*": [ "packages/app-transfer/src/*" ],
-      "@polkadot/app-vanitygen/*": [ "packages/app-vanitygen/src/*" ],
       "@polkadot/apps/*": [ "packages/apps/src/*" ],
       "@polkadot/ui-api/*": [ "packages/ui-api/src/*" ],
       "@polkadot/ui-app/*": [ "packages/ui-app/src/*" ],
